oracle-shared/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "oracle-shared"
version = "0.1.0"
description = "Shared Pydantic contract models for all ORACLE programs"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "google-genai>=1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["oracle_shared"]
